### 2021-10-17 #3830 @whedon
[PRE REVIEW]: sknet: A Python framework for Machine Learning in Complex Networks
**Submitting author:** @TNanukem (<a href="http://orcid.org/0000-0001-5675-554X">Tiago Toledo Junior</a>)
**Repository:** <a href="https://github.com/TNanukem/scikit-net/" target ="_blank">https://github.com/TNanukem/scikit-net/</a>
**Version:** v0.0.1
