### 2021-10-26 46303c2
Added more tests
First step towards 100% coverage. Also, some 'pass' keywords were removed
