+++ sknet/supervised/tests/test_ease_of_access.py
@@ -21,7 +21,7 @@ def X_y_generator():
 
 
 @pytest.fixture
-def module_generator_eigen(X_y_generator):
+def module_generator_eigen_classifier(X_y_generator):
     knn = KNNConstructor(k=3)
     classifier = EaseOfAccessClassifier(t=5)
     classifier.fit(X_y_generator[0], X_y_generator[1], constructor=knn)
@@ -30,7 +30,7 @@ def module_generator_eigen(X_y_generator):
 
 
 @pytest.fixture
-def module_generator_power(X_y_generator):
+def module_generator_power_classifier(X_y_generator):
     knn = KNNConstructor(k=3)
     classifier = EaseOfAccessClassifier(t=5, method='power')
     classifier.fit(X_y_generator[0], X_y_generator[1], constructor=knn)
@@ -39,32 +39,49 @@ def module_generator_power(X_y_generator):
 
 
 @pytest.fixture
-def class_generator(module_generator_eigen,
-                    module_generator_power,
-                    X_y_generator):
+def class_generator_classifier(module_generator_eigen_classifier,
+                               module_generator_power_classifier,
+                               X_y_generator):
 
-    pred_eig = module_generator_eigen.predict(X_y_generator[2])
-    pred_power = module_generator_power.predict(X_y_generator[2])
+    pred_eig = module_generator_eigen_classifier.predict(X_y_generator[2])
+    pred_power = module_generator_power_classifier.predict(X_y_generator[2])
 
-    return module_generator_eigen, module_generator_power, pred_eig, pred_power
-
-
-def test__stationary_distribution(class_generator):
-    np.testing.assert_almost_equal(class_generator[0].P_inf,
-                                   class_generator[1].P_inf)
-
-    pd.testing.assert_frame_equal(class_generator[0].tau_,
-                                  class_generator[1].tau_)
+    return (module_generator_eigen_classifier,
+            module_generator_power_classifier, pred_eig, pred_power)
 
 
-def test_predictions(class_generator):
+def test__stationary_distribution(class_generator_classifier):
+    np.testing.assert_almost_equal(class_generator_classifier[0].P_inf,
+                                   class_generator_classifier[1].P_inf)
+
+    pd.testing.assert_frame_equal(class_generator_classifier[0].tau_,
+                                  class_generator_classifier[1].tau_)
+
+
+def test_predictions(class_generator_classifier):
 
     expected = [1, 0, 2, 1, 1, 0, 1, 2, 1, 1, 2, 0, 0, 0, 0, 1,
                 2, 1, 1, 2, 0, 1, 0, 2, 2, 2, 2, 2, 0, 0, 0, 0,
                 1, 0, 0, 1, 1, 0, 0, 0, 2, 1, 1, 0, 0, 1, 2, 2,
                 1, 2]
 
-    eigen_pred = class_generator[2]
-    power_pred = class_generator[3]
+    eigen_pred = class_generator_classifier[2]
+    power_pred = class_generator_classifier[3]
 
     assert eigen_pred == power_pred == expected
+
+
+def test_raise_on_predict(X_y_generator):
+
+    knn = KNNConstructor(k=3)
+    classifier = EaseOfAccessClassifier(t=5, method='something')
+
+    with pytest.raises(Exception):
+        classifier.fit(X_y_generator[0], X_y_generator[1], constructor=knn)
+        classifier.predict(X_y_generator[2])
+
+
+def test_set_get_params(module_generator_eigen_classifier):
+    classifier = module_generator_eigen_classifier
+    classifier.set_params(t=5)
+    assert classifier.get_params()['t'] == 5
