+++ sknet/supervised/tests/test_ease_of_access.py
@@ -2,15 +2,15 @@ import pytest
 import numpy as np
 import pandas as pd
 
-from sklearn.datasets import load_iris
+from sklearn.datasets import load_iris, load_boston
 from sklearn.model_selection import train_test_split
 
 from sknet.network_construction import KNNConstructor
-from sknet.supervised import EaseOfAccessClassifier
+from sknet.supervised import EaseOfAccessClassifier, EaseOfAccessRegressor
 
 
 @pytest.fixture
-def X_y_generator():
+def X_y_generator_classification():
 
     X, y = load_iris(return_X_y=True)
     X_train, X_test, y_train, y_test = train_test_split(X, y,
@@ -21,36 +21,75 @@ def X_y_generator():
 
 
 @pytest.fixture
-def module_generator_eigen_classifier(X_y_generator):
+def X_y_generator_regression():
+
+    X, y = load_boston(return_X_y=True)
+    X = X[:150, :]
+    y = y[:150]
+    X_train, X_test, y_train, y_test = train_test_split(X, y,
+                                                        test_size=0.33,
+                                                        random_state=42)
+
+    return X_train, y_train, X_test, y_test
+
+
+@pytest.fixture
+def module_generator_eigen_classifier(X_y_generator_classification):
     knn = KNNConstructor(k=3)
     classifier = EaseOfAccessClassifier(t=5)
-    classifier.fit(X_y_generator[0], X_y_generator[1], constructor=knn)
+    classifier.fit(X_y_generator_classification[0],
+                   X_y_generator_classification[1], constructor=knn)
 
     return classifier
 
 
 @pytest.fixture
-def module_generator_power_classifier(X_y_generator):
+def module_generator_power_classifier(X_y_generator_classification):
     knn = KNNConstructor(k=3)
     classifier = EaseOfAccessClassifier(t=5, method='power')
-    classifier.fit(X_y_generator[0], X_y_generator[1], constructor=knn)
+    classifier.fit(X_y_generator_classification[0],
+                   X_y_generator_classification[1], constructor=knn)
 
     return classifier
+
+
+@pytest.fixture
+def module_generator_eigen_regressor(X_y_generator_regression):
+    knn = KNNConstructor(k=3)
+    regressor = EaseOfAccessRegressor(t=5)
+    regressor.fit(X_y_generator_regression[0],
+                  X_y_generator_regression[1], constructor=knn)
+
+    return regressor
+
+
+@pytest.fixture
+def module_generator_power_regressor(X_y_generator_regression):
+    knn = KNNConstructor(k=3)
+    regressor = EaseOfAccessRegressor(t=5, method='power')
+    regressor.fit(X_y_generator_regression[0],
+                  X_y_generator_regression[1], constructor=knn)
+
+    return regressor
 
 
 @pytest.fixture
 def class_generator_classifier(module_generator_eigen_classifier,
                                module_generator_power_classifier,
-                               X_y_generator):
+                               X_y_generator_classification):
 
-    pred_eig = module_generator_eigen_classifier.predict(X_y_generator[2])
-    pred_power = module_generator_power_classifier.predict(X_y_generator[2])
+    pred_eig = module_generator_eigen_classifier.predict(
+        X_y_generator_classification[2]
+    )
+    pred_power = module_generator_power_classifier.predict(
+        X_y_generator_classification[2]
+    )
 
     return (module_generator_eigen_classifier,
             module_generator_power_classifier, pred_eig, pred_power)
 
 
-def test__stationary_distribution(class_generator_classifier):
+def test__stationary_distribution_classifier(class_generator_classifier):
[…]
