### 2021-11-01 #3864 @dvysardana
Hello @osorensen

Yes, I have completed my review. The statement of need needs to be updated
to clearly mention the target audience. Further, the paper lacks a
description and comparison of other similar packages that already exist in
the community.

Can you please confirm that all the review requirements are complete now?

Thanks
Divya Sardana

On Sun, Oct 31, 2021 at 12:01 AM Øystein Sørensen ***@***.***>
wrote:

> Thanks @dvysardana <https://github.com/dvysardana>. Do I understand
> correctly that you think a statement of need is missing but otherwise
> everything is fine?
>
