+++ paper/paper.bib
@@ -74,3 +74,14 @@ volume = {23},
 journal = {IEEE Transactions on Neural Networks},
 doi = {10.1109/TNNLS.2011.2181866}
 }
+
+@article{giotto-tda,
+  author  = {Guillaume Tauzin and Umberto Lupo and Lewis Tunstall and Julian Burella P\'{e}rez and Matteo Caorsi and Anibal M. Medina-Mardones and Alberto Dassatti and Kathryn Hess},
+  title   = {giotto-tda: A Topological Data Analysis Toolkit for Machine Learning and Data Exploration},
+  journal = {Journal of Machine Learning Research},
+  year    = {2021},
+  volume  = {22},
+  number  = {39},
+  pages   = {1-6},
+  url     = {http://jmlr.org/papers/v22/20-325.html}
+}
