### 2021-11-02 #3864 @TNanukem
Hello @dvysardana

I've added more information about the target audience on the README and on the sphinx documentation.
Also, I added more information about other packages on the paper.

Do those changes meet the criteria for your revision? If not, do you have any suggestions on how to improve it to make it acceptable?

Thank you for your review!
