### 2021-11-02 #3864 @dvysardana
Hello @tnanukem

Thanks for adding the additional information. These changes look good to
me. I have no other suggestions for improvement.

Best Regards
Divya Sardana

On Tue, Nov 2, 2021 at 4:50 AM Tiago Toledo Junior ***@***.***>
wrote:

> Hello @dvysardana <https://github.com/dvysardana>
>
> I've added more information about the target audience on the README and on
> the sphinx documentation.
> Also, I added more information about other packages on the paper.
>
> Do those changes meet the criteria for your revision? If not, do you have
> any suggestions on how to improve it to make it acceptable?
>
> Thank you for your review!
>
