### 2021-11-15 76abd79
Aggregation method on modulatiry label propagator
Added the aggregation method proposed by Silva et al. for the modularity label propagation algorithm
