+++ changelog.md
@@ -1,0 +1,21 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.1.0] - xxxx-xx-xx
+
+### Added
+- Modularity Label Propagation now allows a network reduction proposed by Silva et al.
+
+### Changed
+
+### Fixed
+
+### Removed
+
+### Deprecated
+
+[0.0.1]: https://github.com/TNanukem/scikit-net/releases/tag/v0.0.1
