### 2021-11-17 #3864 @osorensen
@drj11, how is it going with your review? If there are issues which you think the authors should address, please mention them in this thread or by opening an issue in the [source repository](https://github.com/TNanukem/scikit-net/).
