### 2021-12-01 626d943
Added single linkage heuristic constructor
