+++ sknet/network_construction/dataset_constructors.py
@@ -3,6 +3,7 @@ import pandas as pd
 import networkx as nx
 
 from abc import ABCMeta, abstractmethod
+from sklearn.metrics import pairwise_distances
 from sklearn.neighbors import KDTree, BallTree
 
 
@@ -182,7 +183,7 @@ class KNNConstructor(BaseConstructor):
     >>> knn_c = KNNConstructor(k=3)
     >>> knn_c.fit(X, y)
     >>> G = knn_c.transform()
-    >>> print(len(G.nodes))
+    >>> # print(len(G.nodes))
     150
 
     Notes
@@ -327,7 +328,7 @@ class EpsilonRadiusConstructor(BaseConstructor):
     >>> eps_c = EpsilonRadiusConstructor(epsilon=3)
     >>> eps_c.fit(X, y)
     >>> G = eps_c.transform()
-    >>> print(len(G.nodes))
+    >>> # print(len(G.nodes))
     150
 
     Notes
@@ -482,7 +483,7 @@ class KNNEpislonRadiusConstructor(BaseConstructor):
     >>> ke_c = KNNEpislonRadiusConstructor(k=3, epsilon=0.3)
     >>> ke_c.fit(X, y)
     >>> G = ke_c.transform()
-    >>> print(len(G.nodes))
+    >>> # print(len(G.nodes))
     150
 
     Notes
@@ -650,3 +651,195 @@ def _tree_selector(X, leaf_size=40, metric='minkowski'):
     # High dimensional spaces are fit to Ball Tree
     if X.shape[1] >= 30:
         return BallTree(X, leaf_size=leaf_size, metric=metric)
+
+
+class SingleLinkageHeuristicConstructor(BaseConstructor):
+    """
+    Use Single Linkage Heuristics to generate a complex network from
+    tabular data
+
+    Parameters
+    ----------
+    k : int, default=3
+        The number of closests points between two grops to be considered
+        to create an edge.
+    lambda_ : positive float, default=0.1
+        Multiplying factor on the average dissimilarity on the groups to
+        define the critical distance
+    sep_comp : boolean, default=True
+        If True and if y is not None, then each class of the dataset
+        will be a separated component, so nodes from one class will only
+        be connected to those of the same class. If False then this
+        restriction is not applied.
+    metric : str or DistanceMetric object, default='euclidean'
+        The distance metric to use for the neighborhood tree. Refer
+        to the DistanceMetric class documentation from sklearn for a list
+        of available metrics
+    n_jobs : int, default=None
+        The number of parallel jobs to run for neighbors search.
+        None means 1 unless in a joblib.parallel_backend context and -1 means
+        using all processors.
+
+    Examples
+    --------
+    >>> from sklearn.datasets import load_iris
+    >>> from dataset_constructors import SingleLinkageHeuristicConstructor
+    >>> X, y = load_iris(return_X_y = True)
+    >>> ch = SingleLinkageHeuristicConstructor(k=3, epsilon=0.3)
+    >>> ch.fit(X, y)
+    >>> G = ke_c.transform()
+    >>> # print(len(G.nodes))
+    150
+
+    References
+    ----------
+    Cupertino, T.H., Huertas, J., & Zhao, L. (2013). Data clustering using
+    controlled consensus in complex networks. Neurocomputing, 118, 132-140.
+
+    """
+    def __init__(self, k=3, lambda_=0.1, sep_comp=False,
+                 metric='euclidean', n_jobs=None):
+        self.k = k
+        self.lambda_ = lambda_
+        self.sep_comp = sep_comp
+        self.metric = metric
+        self.n_jobs = n_jobs
+
+    def get_params(self, deep=True):
+        return {'k': self.k, 'lambda_': self.lambda_,
[…]
