### 2021-12-07 81c528c
Recurence network constructor added
