+++ sknet/network_construction/__init__.py
@@ -5,3 +5,4 @@ from .dataset_constructors import KNNEpislonRadiusConstructor
 from .dataset_constructors import SingleLinkageHeuristicConstructor
 from .time_series_constructors import UnivariateCorrelationConstructor
 from .time_series_constructors import MultivariateCorrelationConstructor
+from .time_series_constructors import UnivariateRecurrenceNetworkConstructor
