### 2021-12-07 #3864 @osorensen
@imw, I have added a review checklist for you at the top of this thread, which you can use for your review. You find the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Please let me know if you need any assistance in the process.
