### 2021-12-08 #3864 @osorensen
Thanks for getting back to us @drj11. I've tried to ping you a few times, but I completely understand that it's easy to loose track of e-mails. It would be very much appreciated if you could complete your review, even though we have already secured a third reviewer. Having three reviewers rather than two is a good thing.

And thanks for your comment regarding deadlines. However, this might be slightly problematic to implement, since the reviews in JOSS are supposed to be interactive. Sometimes, when the reviewers have a large number of suggestions, it can take many months for the submitting authors to fix the issues proposed by the reviewers, so it's hard to set deadlines for reviews.
