### 2021-12-08 TNanukem/scikit-net#34 @drj11
auxiliar: minor typo
[I'm completing the JOSS review at https://github.com/openjournals/joss-reviews/issues/3864#issuecomment-988725114]

This typo will not affect the review:

the word _auxiliar_ in Page 2 of https://raw.githubusercontent.com/openjournals/joss-papers/joss.03864/joss.03864/10.21105.joss.03864.pdf should be _auxiliary_ .
