### 2021-12-08 TNanukem/scikit-net#35 @drj11
Snippet in README is broken
If you try and use the code in the [snippet in the README](https://github.com/TNanukem/scikit-net#high_brightness-quickstart) an `Exception` is raised because `train_test_split` is not defined.

```
>>> from sklearn.datasets import load_iris
>>> from sknet.network_construction import KNNConstructor
>>> from sknet.supervised import EaseOfAccessClassifier
>>>
>>> X, y = load_iris(return_X_y = True)
>>> X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'train_test_split' is not defined
```

(using current tip of main 914a3f9aa138a659361b0ea3b658e472170ae12a)

[later: the same function is also referenced several times in the documentation]
