### 2021-12-08 TNanukem/scikit-net#34 @drj11
The same typo appears in https://tnanukem.github.io/scikit-net/main/user_guide/index.html as well

Because _auxiliar_ is a noun and in English any noun can appear in adjectival position, it's not technically incorrect. But expert English writers would probably expect to see _auxiliary_ being used.
