### 2021-12-08 TNanukem/scikit-net#37 @drj11
Contributing guidelines are missing
The JOSS review checklist asks that:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Because the project is hosted on GitHub it is at least implicitly clear that opening an issue can be used for 2) and 3).

There are no guidelines for how to contribute to the software (or, if there are, they are not made clear).

I would suggest a `contributing.md` file, which could also cover 2) and 3) more explicitly.

This is blocking review, I won't check the JOSS review checkbox until this is resolved.
