### 2021-12-08 TNanukem/scikit-net#35 @drj11
I would like to consider that this snippet in the README covers the JOSS checklist item "Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)."

Technically the JOSS wording does not require that the example _work_, however i think it would show willing to get this example working before i check the box.
