### 2021-12-08 #3864 @drj11
@TNanukem I have opened a few issues, but only two block the review:

https://github.com/TNanukem/scikit-net/issues/35

https://github.com/TNanukem/scikit-net/issues/37

I think that these should be easy to resolve.

In the written paper I particularly appreciated the Statement of Need and the catalogue of implemented algorithms.
