### 2021-12-10 TNanukem/scikit-net#37 @imw
@drj11 Do you think the [development](https://tnanukem.github.io/scikit-net/main/development/index.html) page of the docs is sufficient? Or would you like to see this reflected in the code itself?
