### 2021-12-10 TNanukem/scikit-net#35 @imw
Looks like the quickstart is missing:
```
from sklearn.model_selection import train_test_split
from sklean.metrics import accuracy_score
```
