### 2021-12-10 TNanukem/scikit-net#39 @imw
Data transformation example in docs doesn't work
Initial data transformation example has:
```
from dataset_constructors import KNNEpislonRadiusConstructor
```
This should probably be:
```
from sknet.network_construction.dataset_constructors import KNNEpislonRadiusConstructor
```
