### 2021-12-10 TNanukem/scikit-net#40 @imw
Minor copy edits
There are a handful of minor grammatical issues in the JOSS paper. Rather than list them here, it seemed easier to open a PR:
https://github.com/TNanukem/scikit-net/compare/docs/joss_paper...imw:patch-1
