### 2021-12-10 #3864 @imw
@TNanukem I also opened a couple of issues.:
https://github.com/TNanukem/scikit-net/issues/39
https://github.com/TNanukem/scikit-net/issues/40
Also quite minor. Nice library!
