### 2021-12-10 TNanukem/scikit-net#37 @drj11
aha, i didn't actually find that. I still think a `contributing.md` would be a good idea. Either with a brief summary and a link to that page, or the text of that page.

If you create a `contributing.md` GitHub creates a prompt and link when you open a new issue. It's a becoming a standard practice.
