### 2021-12-11 f3fc3bd
Several minor corrections to documentation
Corrected documentation problems from the JOSS reviewing process
