+++ docs/source/getting_started/installation.rst
@@ -1,7 +1,8 @@
 Installation
 ============
 
-In order to install the sknet you can use pip using the ``scikit-net`` name.
+In order to install the sknet you can use pip using the ``scikit-net`` name as follows:
 
 .. code-block:: python
+
     pip install scikit-net
