### 2021-12-11 bae6318
Changed the type of code on installation
Changed from python to console
