+++ docs/source/getting_started/installation.rst
@@ -3,6 +3,6 @@ Installation
 
 In order to install the sknet you can use pip using the ``scikit-net`` name as follows:
 
-.. code-block:: python
+.. code-block:: console
 
     pip install scikit-net
