### 2021-12-11 TNanukem/scikit-net#40 @TNanukem
I created the pull request and merged it on the branch. Thank you!
