### 2021-12-11 #3864 @TNanukem
Hello @drj11 and @imw

First of all, thank you for your review!

I think I solved most of the issues (all the blocking ones at least). Two issues remain open but, as they are non-blocking, I will get to them next week.

If there is any other point I will be happy to work on it. Thanks.
