### 2021-12-11 #3864 @TNanukem
Hello @imw and @drj11

Thank you for your review! I think I solved most of the issues.

Only the one related to the warning on the unit tests is still open. Since it will change the expected value of the test, I will tackle it after. However, I think the blocking issues are all solved.

If there is any additional point I will be happy to work on it.

Have a good weekend everyone!
