### 2021-12-14 #3864 @osorensen
@imw, when you have time, please let us know if the two Issues you opened in the [source repository](https://github.com/TNanukem/scikit-net) have been properly addressed.
