### 2021-12-14 TNanukem/scikit-net#44 @osorensen
Editorial comments on manuscript
Hello @TNanukem

Here are some editorial comments related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/3864).

I suggest you make the following small changes in the paper:

- On line 8 in the paper, change from "the graph theory" to just "graph theory".
- Make sure Python is spelled consistently. For example, on line 18 "Python" is used, and on line 26 "python" is used.
- Line 34, change from "the `scikit-learn`" to just "`scikit-learn`".
- Line 35, change from "it main focus" to "its main focus".
- Line 37, change from "on their data" to "in their data".
- Line 43, change from "A `unsupervised`" to "An `unsupervised`".
- Lines 45-46, change from "where one have" to "where one has".
