### 2021-12-15 83c70df
Initial Constructor Structure
