### 2021-12-15 TNanukem/scikit-net#44 @TNanukem
Thank you for your comments @osorensen. Solved them on my last commit to the branch.
