### 2021-12-15 #3864 @TNanukem
Hello @osorensen

Thank you for your comments, I solved them. As for the other requests:
- Release v0.0.1
- DOI: 10.5281/zenodo.5783027
- Here's the link: https://zenodo.org/record/5783027
